handle camera xscale flip & improve UI

+++ app/components/MoneyRecord.tsx
@@ -61,19 +61,21 @@ function MoneyRecord(props) {
     }, 0);
 
     async function handleDelete(event) {
-        const {name} = event.target;
-        const response = await fetch('/api/money-record', {
-            method: 'DELETE',
-            body: name
-        });
-        setRecords(prevRecords => prevRecords.filter(record => record._id != name))
-        event.preventDefault();
+        if (confirm('Are you sure to delete that item?')) {
+            const {name} = event.target;
+            const response = await fetch('/api/money-record', {
+                method: 'DELETE',
+                body: name
+            });
+            setRecords(prevRecords => prevRecords.filter(record => record._id != name))
+            event.preventDefault();
+        }
     }
 
     return (
     <div style={{width:'90vw', height:'auto', backgroundColor:'black', opacity: '0.8', borderRadius: '15px'}}>
         <h2>Money Record</h2>
-        {records.slice(0,5).map(record => ( // Limit to the first 5 items
+        {records.slice(-5).map(record => ( // Limit to the last 5 items
             <div key={record._id} style={{display:'flex', margin: '10px 0', padding: '5px'}}>
                 <h3 style={{width:'100%'}}>{record.date} ${record.value} {record.description}</h3>
                 <button onClick={handleDelete} name={record._id} style={{marginLeft:'10px'}}>delete</button>
@@ -81,7 +83,7 @@ function MoneyRecord(props) {
         ))}
         <h3>This month spent: {sum}</h3>
         <form style={{display:'flex', margin: '5px 0', padding: '5px'}}>
-            <input id="valueInput" style={{width:'100%', marginRight: '5px'}} onChange={handleMoneyChange} name="money" value={money}></input>
+            <input id="valueInput" style={{width:'100%', marginRight: '5px'}} type="tel" onChange={handleMoneyChange} name="money" value={money}></input>
             <input id="descriptionInput" name="description" onChange={handleDescriptionChange} value={description}></input>
             <button style={{marginLeft:'10px'}} onClick={handleData}>Submit</button>
         </form>
